+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## CHANGELOG
 
+### 0.1.2: brighter colors
+- made most colors a little brighter
+- added a test for generic logging
+
 ### 0.1.1: initial commit
 - initial commit